move functions using websockets to their own files
