+++ server/src/lib.rs
@@ -1,8 +1,9 @@
 pub mod config;
-pub mod dataio;
-pub mod errors;
 pub mod webcast;
-pub mod dbupdate;
+mod dataio;
+mod errors;
+mod dbupdate;
+mod websockets;
 
 use data::configdata::ConfigContest;
 
@@ -21,9 +22,6 @@ use std::sync::Arc;
 
 use tokio::sync::broadcast;
 use tokio::sync::Mutex;
-
-use futures::{SinkExt, StreamExt, stream::SplitSink};
-use warp::ws::Message;
 
 use warp::Filter;
 
@@ -46,12 +44,12 @@ pub fn route_contest_public_data(
         .and(warp::ws())
         .and(with_db(shared_db.clone()))
         .and(warp::any().map(move || tx.subscribe()))
-        .map(|ws: warp::ws::Ws, db, rx| ws.on_upgrade(move |ws| serve_all_runs_ws(ws, db, rx)));
+        .map(|ws: warp::ws::Ws, db, rx| ws.on_upgrade(move |ws| websockets::serve_all_runs_ws(ws, db, rx)));
 
     let timer = warp::path("timer")
         .and(warp::ws())
         .and(with_db(shared_db.clone()))
-        .map(|ws: warp::ws::Ws, db| ws.on_upgrade(move |ws| serve_timer_ws(ws, db)));
+        .map(|ws: warp::ws::Ws, db| ws.on_upgrade(move |ws| websockets::serve_timer_ws(ws, db)));
 
     let contest_file = warp::path("contest")
         .and(with_db(shared_db.clone()))
@@ -87,68 +85,6 @@ pub fn serve_urlbase(
 async fn serve_runs(runs: Arc<Mutex<DB>>) -> Result<impl warp::Reply, warp::Rejection> {
     let db = runs.lock().await;
     Ok(serde_json::to_string(&*db.latest()).map_err(SerializationError)?)
-}
-
-async fn serve_timer_ws(ws: warp::ws::WebSocket, runs: Arc<Mutex<DB>>) {
-    let (mut tx, _) = ws.split();
-
-    let fut = async move {
-        let dur = tokio::time::Duration::new(1, 0);
-        let mut interval = tokio::time::interval(dur);
-        let mut old = data::TimerData::fake();
-
-        loop {
-            interval.tick().await;
-            let l = runs.lock().await.timer_data();
-
-            if l != old {
-                old = l;
-                let message = serde_json::to_string(&l).map(Message::text)
-                    .unwrap_or_else(|error| panic!("Could not convert `{:?}' to a message: {:?}", l, error));
-                tx.send(message).await.unwrap_or_else(|error| panic!("Could not send message: {:?}", error));
-            }
-        }
-    };
-
-    tokio::task::spawn(fut);
-}
-
-async fn convert_and_send(tx: &mut SplitSink<warp::ws::WebSocket, Message>, r : data::RunTuple) -> bool {
-    match serde_json::to_string(&r).map(Message::text) {
-        Err(e) => {
-            panic!("Error preparing run {:?} message: {:?}", r, e);
-        }
-        Ok(m) => tx.send(m).await.is_ok()
-    }
-}
-
-async fn serve_all_runs_ws(
-    ws: warp::ws::WebSocket,
-    runs: Arc<Mutex<DB>>,
-    mut rx: broadcast::Receiver<data::RunTuple>,
-) {
-    let (mut tx, _) = ws.split();
-
-    let fut = async move {
-        {
-            let lock = runs.lock().await;
-
-            for r in lock.all_runs() {
-                if !convert_and_send(&mut tx, r).await {
-                    return
-                }
-            }
-        }
-
-        loop {
-            let r = rx.recv().await.expect("Expected a RunTuple");
-            if !convert_and_send(&mut tx, r).await {
-                return
-            }
-        }
-    };
-
-    tokio::task::spawn(fut);
 }
 
 async fn serve_all_runs_secret(runs: Arc<Mutex<DB>>) -> Result<impl warp::Reply, warp::Rejection> {
